+++ changelog.md
@@ -2,7 +2,8 @@
 
 ## dev
 
-- Fixed Form not automatically syncing inputs when pressing primary action button.
+Fixed Form not automatically syncing inputs when pressing primary action button.
+Added support for Tags `options.tagClass`.
 
 ## 8.2.3 (master)